+++ server/nmtwizard/common.py
@@ -364,7 +364,8 @@ def cmd_docker_run(lxpu, docker_options, task_id,
 
         for arg in docker_command:
             if arg.startswith('${TMP_DIR}'):
-                arg = '/root/tmp/%s%s' % (task_id, arg[10:])
+                # when there are multi-files, use replace instead of concatenate
+                arg = arg.replace('${TMP_DIR}', '/root/tmp/%s' % task_id)
             cmd += '_o_' + arg.replace("'", "'\"'\"'")
 
         return cmd.replace("\n", "\\\\n").replace("_o_", "\n"), str(env).replace("'", '"')